Add files via upload
There are not many changes since your version of the code.

Preview button works for both ORCA and GAUSSIAN submissions (I think). Molecular coordinates are omitted from the script so the readability of the key info is clearer, however they are still in the input file of course and that still works fine.

I also implemented a script that can be used on CSF3 (kind of). Right now it only works for ORCA, however should be able to make it work for gaussian in the same way, I was just lazy.

I also changed your QREP File slightly, It now accepts files from any directory n exports them to any directory fine (from what I tested anyway). If no directory is selected, then a QCALC directory is created which stores all the data. I was thinking maybe we could make this the default directory stuff is exported too, however might be a good idea to still keep a choice to export to any directory.

Any changes u can just lmk tomorrow